fixup long funcs to return Result instead of Result<Option>
