+++ crates/hyperdrive-math/src/long/open.rs
@@ -1,5 +1,5 @@
 use eyre::{eyre, Result};
-use fixed_point::FixedPoint;
+use fixed_point::{fixed, FixedPoint};
 
 use crate::{calculate_rate_given_fixed_price, State, YieldSpace};
 
@@ -42,6 +42,67 @@ impl State {
         }
 
         Ok(bond_amount - self.open_long_curve_fee(base_amount)?)
+    }
+
+    /// Calculates the derivative of
+    /// [calculate open long](State::calculate_open_long) with respect to the
+    /// base amount.
+    ///
+    /// We calculate the derivative of the long amount `$y(x)$` as:
+    ///
+    /// ```math
+    /// y'(x) = y_{*}'(x) - c'(x)
+    /// ```
+    ///
+    /// Where `$y_{*}'(x)$` is the derivative of `$y_{*}(x)$` and `$c^{\prime}(x)$`
+    /// is the derivative of `$c(x)$`, the [long curve fee](State::open_long_curve_fee).
+    /// `$y_{*}^{\prime}(x)$` is given by:
+    ///
+    /// ```math
+    /// y_{*}'(x) = \left( \mu \cdot (z + \tfrac{x}{c}) \right)^{-t_s}
+    ///             \left(
+    ///                 k - \tfrac{c}{\mu} \cdot
+    ///                 \left(
+    ///                     \mu \cdot (z + \tfrac{x}{c}
+    ///                 \right)^{1 - t_s}
+    ///             \right)^{\tfrac{t_s}{1 - t_s}}
+    /// ```
+    ///
+    /// and `$c^{\prime}(x)$` is given by:
+    ///
+    /// ```math
+    /// c^{\prime}(x) = \phi_{c} \cdot \left( \tfrac{1}{p} - 1 \right)
+    /// ```
+    pub(super) fn calculate_open_long_derivative(
+        &self,
+        base_amount: FixedPoint,
+    ) -> Result<FixedPoint> {
+        let share_amount = base_amount / self.vault_share_price();
+        let inner =
+            self.initial_vault_share_price() * (self.effective_share_reserves()? + share_amount);
+        let mut derivative = fixed!(1e18) / (inner).pow(self.time_stretch())?;
+
+        // It's possible that k is slightly larger than the rhs in the inner
+        // calculation. If this happens, we are close to the root, and we short
+        // circuit.
+        let k = self.k_down()?;
+        let rhs = self.vault_share_price().mul_div_down(
+            inner.pow(self.time_stretch())?,
+            self.initial_vault_share_price(),
+        );
+        if k < rhs {
+            return Err(eyre!("Open long derivative is undefined."));
+        }
+        derivative *= (k - rhs).pow(
+            self.time_stretch()
+                .div_up(fixed!(1e18) - self.time_stretch()),
+        )?;
+
+        // Finish computing the derivative.
+        derivative -=
+            self.curve_fee() * ((fixed!(1e18) / self.calculate_spot_price()?) - fixed!(1e18));
+
+        Ok(derivative)
     }
 
     /// Calculate an updated pool state after opening a long.
@@ -126,7 +187,10 @@ mod tests {
 
     use ethers::types::{I256, U256};
     use fixed_point::fixed;
-    use hyperdrive_test_utils::{chain::TestChain, constants::FUZZ_RUNS};
+    use hyperdrive_test_utils::{
+        chain::TestChain,
+        constants::{FAST_FUZZ_RUNS, FUZZ_RUNS},
+    };
     use hyperdrive_wrappers::wrappers::ihyperdrive::Options;
     use rand::{thread_rng, Rng, SeedableRng};
     use rand_chacha::ChaCha8Rng;
@@ -420,6 +484,66 @@ mod tests {
             alice.reset(Default::default()).await?;
             bob.reset(Default::default()).await?;
             celine.reset(Default::default()).await?;
+        }
+
+        Ok(())
+    }
+
+    /// This test empirically tests the derivative returned by
+    /// `calculate_open_long_derivative` by calling `calculate_open_long` at two
+    /// points and comparing the empirical result with the output of
+    /// `calculate_open_long_derivative`.
+    #[tokio::test]
+    async fn fuzz_open_long_derivative() -> Result<()> {
+        let mut rng = thread_rng();
+        // We use a relatively large epsilon here due to the underlying fixed point pow
+        // function not being monotonically increasing.
+        let empirical_derivative_epsilon = fixed!(1e12);
+        // TODO pretty big comparison epsilon here
+        let test_comparison_epsilon = fixed!(10e18);
+
+        for _ in 0..*FAST_FUZZ_RUNS {
+            let state = rng.gen::<State>();
+            let amount = rng.gen_range(fixed!(10e18)..=fixed!(10_000_000e18));
+
+            // We need to catch panics here because FixedPoint panics on overflow or underflow.
+            let f_x = match panic::catch_unwind(|| state.calculate_open_long(amount)) {
+                Ok(result) => match result {
+                    Ok(result) => result,
+                    Err(_) => continue, // Err; the amount results in the pool being insolvent.
+                },
+                Err(_) => continue, // panic; likely in FixedPoint
+            };
+
+            let f_x_plus_delta = match panic::catch_unwind(|| {
+                state.calculate_open_long(amount + empirical_derivative_epsilon)
+            }) {
+                Ok(result) => match result {
+                    Ok(result) => result,
+                    Err(_) => continue,
+                },
+                // If the amount results in the pool being insolvent, skip this iteration.
+                Err(_) => continue,
+            };
+            // Sanity check.
+            assert!(f_x_plus_delta > f_x);
+
+            let empirical_derivative = (f_x_plus_delta - f_x) / empirical_derivative_epsilon;
+            let open_long_derivative = state.calculate_open_long_derivative(amount)?;
+            let derivative_diff = if open_long_derivative >= empirical_derivative {
+                open_long_derivative - empirical_derivative
+            } else {
+                empirical_derivative - open_long_derivative
+            };
+            assert!(
+                derivative_diff < test_comparison_epsilon,
+                "expected (derivative_diff={}) < (test_comparison_epsilon={}), \
+                calculated_derivative={}, emperical_derivative={}",
+                derivative_diff,
+                test_comparison_epsilon,
+                open_long_derivative,
+                empirical_derivative
+            );
         }
 
         Ok(())
