+++ crates/hyperdrive-math/src/long/targeted.rs
@@ -305,10 +305,7 @@ impl State {
 
         // b'(x) = -y'(x)
         // -b'(x) = y'(x)
-        let long_amount_derivative = match self.calculate_open_long_derivative(base_amount)? {
-            Some(derivative) => derivative,
-            None => return Err(eyre!("long_amount_derivative failure.")),
-        };
+        let long_amount_derivative = self.calculate_open_long_derivative(base_amount)?;
 
         // v(x) = a(x) / b(x)
         // v'(x) = ( b(x) * a'(x) - a(x) * b'(x) ) / b(x)^2
